requirements.txt
numpy
pandas
scikit-learn
matplotlib
pillow
torch
torchvision
joblib
gradio
tqdm
umap-learn